fix some invalid code
