+++ src/interfaces/cVoxelType.ts
@@ -94,464 +94,3 @@ export const cVoxelTest: CVoxelVisType[] = [
   },
 ];
 
-export const initCVoxel: CVoxel[] = [
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "1", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "3", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "3", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "3", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "1.6", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "1.5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.9", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "0.9", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "2", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.8", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "3", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "1.5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "5", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "",
-    value: "15", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "aaaa",
-    fromSig: "bbbbb",
-  },
-  {
-    to: "", // payee address
-    from: "", // payer address
-    summary: "", // work summary
-    deliverable: "aaaa",
-    value: "30", // reward value
-    tokenSymbol: "eth", // eth, usdc, etc
-    tokenDecimal: 18,
-    networkId: 0, // eth mainnet = 1
-    issuedTimestamp: "", //block timestamp
-    txHash: "",
-    tags: [], //tags
-    genre: "",
-    jobType: "FullTime", // default=OneTime
-    toSig: "",
-    fromSig: "bbbbb",
-  },
-];
